requirement.txt
streamlit
google-generativeai
python-dotenv
youtube_tramscript_api
pathlib